requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9.0,<1.0
Pillow
scikit-learn
numpy
pandas
requests
